+++ packages/EventMap/src/EventMap.spec.ts
@@ -1,7 +1,8 @@
 
 import { describe, it, expect } from "@jest/globals";
-import { html, TemplateResult, render } from "lit-html";
-import { EventMap, EventMapHandlerInfo, eventsListenAt, event } from "./EventMap";
+import { html, fixture, FixtureElement } from "@domx/testutils";
+import { EventMap, EventMapHandlerInfo } from "./EventMap";
+import { eventsListenAt, event } from "./decorators";
 
 
 const test1Html = html`
@@ -21,22 +22,6 @@ const defaultListenAtHtml = html`
 
 let __firedEvents: any = {};
 
-interface RestorableElement extends HTMLElement {
-    restore: Function
-};
-
-function fixture(html:TemplateResult): RestorableElement {
-  let fixture = document.createElement("div");
-  fixture.setAttribute("fixture", "");
-  document.body.appendChild(fixture);
-
-  render(html, fixture);
-  const el = fixture.firstElementChild as RestorableElement;
-
-  // set the remove method to remove the fixture
-  el.restore = () => fixture.remove()
-  return el;
-}
 
 
 describe("EventMap", function () {
@@ -86,11 +71,11 @@ describe("EventMap", function () {
   });
 
   describe("Events", function () {
-    var frag:RestorableElement, child:Element, event:CustomEvent;
+    var frag:FixtureElement, child:Element, event:CustomEvent;
 
     beforeEach(function () {
       frag = fixture(test1Html);
-      child = frag.querySelector("#child") as RestorableElement;
+      child = frag.querySelector("#child") as FixtureElement;
       event = new CustomEvent('event-to-trigger', { bubbles: true, composed: true });      
     });
 
@@ -118,10 +103,23 @@ describe("EventMap", function () {
       expect(() => child.dispatchEvent(event)).not.toThrow();
       window.removeEventListener("event-to-trigger", listener);
     });
+
+    it("does not stop event propagation set to false", function () {
+      let listenerTriggered = false;
+      let listener = () => {
+        listenerTriggered = true
+      };
+
+      const event = new CustomEvent('event-to-test-stop-propagation', { bubbles: true, composed: true });
+      window.addEventListener("event-to-test-stop-propagation", listener);
+      child.dispatchEvent(event)
+      expect(listenerTriggered).toBe(true);
+      window.removeEventListener("event-to-test-stop-propagation", listener);
+    });
   });
 
   describe("Events at parent", function () {
-    var frag:RestorableElement, child:Element, event:CustomEvent;
+    var frag:FixtureElement, child:Element, event:CustomEvent;
 
     beforeEach(function () {
       frag = fixture(test1Html);
@@ -156,7 +154,7 @@ describe("EventMap", function () {
   });
 
   describe("Events at window", function () {
-    var frag:RestorableElement, event:CustomEvent;
+    var frag:FixtureElement, event:CustomEvent;
 
     beforeEach(function () {
       frag = fixture(test1Html);
@@ -180,7 +178,7 @@ describe("EventMap", function () {
   });
 
   describe("Mixed in events", () => {
-    let el:RestorableElement;
+    let el:FixtureElement;
 
     beforeEach(() => {
       el = fixture(html`<element-with-mixed-in-events></element-with-mixed-in-events>`);
@@ -207,7 +205,7 @@ describe("EventMap", function () {
   });
 
   describe("eventsListenAt", () => {
-    let frag:RestorableElement;
+    let frag:FixtureElement;
 
     beforeEach(() => {
       frag = fixture(defaultListenAtHtml);
@@ -237,12 +235,23 @@ describe("EventMap", function () {
         expect(__firedEvents._triggerOnDefault).toBe(true);
         el.restore();
       });
+
+      it("can stop immediate propagation", () => {
+        const el = fixture(html`<events-listen-at-descriptor></events-listen-at-descriptor>`);
+        const event = new CustomEvent("trigger-on-default", { bubbles: true, composed: true });
+        const spy = jest.spyOn(event, "stopImmediatePropagation");
+        window.dispatchEvent(event);
+        expect(spy).toBeCalled();
+        spy.mockRestore();
+        el.restore();
+      });
     });
 
     describe("event", () => {
       it("can be used on a method to define an event handler", () => {
         const el = fixture(html`<event-descriptor></event-descriptor>`);
-        window.dispatchEvent(new CustomEvent("trigger-on-default", { bubbles: true, composed: true }));
+        const event = new CustomEvent("trigger-on-default", { bubbles: true, composed: true });
+        window.dispatchEvent(event);
         expect(__firedEvents._triggerOnDefault).toBe(true);
         el.restore();
       });
@@ -251,6 +260,16 @@ describe("EventMap", function () {
         const el = fixture(html`<event-descriptor></event-descriptor>`);
         el.dispatchEvent(new CustomEvent("trigger-on-listen-at", { bubbles: true, composed: true }));
         expect(__firedEvents._triggerOnListenAt).toBe(true);
+        el.restore();
+      });
+
+      it("can stop immediate propagation", () => {
+        const el = fixture(html`<event-descriptor></event-descriptor>`);
+        const event = new CustomEvent("trigger-on-listen-at", { bubbles: true, composed: true });
+        const spy = jest.spyOn(event, "stopImmediatePropagation");
+        el.dispatchEvent(event);
+        expect(spy).toBeCalled();
+        spy.mockRestore();
         el.restore();
       });
     });
@@ -266,7 +285,7 @@ describe("EventMap", function () {
         });
   
         const frag = fixture(test1Html);
-        const child = frag.querySelector("#child") as RestorableElement;
+        const child = frag.querySelector("#child") as FixtureElement;
         const event = new CustomEvent('event-to-trigger', { bubbles: true, composed: true });      
         expect(__firedEvents._toTrigger).toBeUndefined();
         child.dispatchEvent(event);
@@ -309,6 +328,10 @@ describe("EventMap", function () {
       expect(__firedEvents.handler1Count).toBe(2);
     });
   });
+
+
+
+  // add test for event and eventsListenAt decorators
 });
 
 
@@ -325,16 +348,24 @@ class EventMapTest1 extends EventMap(HTMLElement) {
     "event-to-test-window": {
       listenAt: "window",
       handler: "_toTestWindow"
+    },
+    "event-to-test-stop-propagation": {
+      listenAt: "parent",
+      handler: "_testStopPropagation",
+      stopPropagation: false
     }
   };
   _toTrigger() {
       __firedEvents._toTrigger = "triggered";
   }
   _toTestParent() {
-      __firedEvents._toTrigger = "triggeredOnParent";
+    __firedEvents._toTrigger = "triggeredOnParent";
   }
   _toTestWindow() {
-      __firedEvents._toTrigger = "triggeredOnWindow";
+    __firedEvents._toTrigger = "triggeredOnWindow";
+  }
+  _testStopPropagation() {
+    __firedEvents._toTrigger = "testStopPropagation";
   }
 }
 customElements.define("event-map-test-1", EventMapTest1);
@@ -433,7 +464,9 @@ class DefaultListenAt extends EventMap(HTMLElement) {
 customElements.define("default-listen-at", DefaultListenAt);
 
 
-@eventsListenAt("window")
+@eventsListenAt("window", {
+  stopImmediatePropagation: true
+})
 class EventsListenAtDescriptor extends EventMap(HTMLElement) {
   static events = {
       "trigger-on-default": "_triggerOnDefault",
@@ -453,7 +486,7 @@ class EventDescriptor extends EventMap(HTMLElement) {
       __firedEvents._triggerOnDefault = true;
   }
 
-  @event("trigger-on-listen-at", {listenAt: "parent"})
+  @event("trigger-on-listen-at", {listenAt: "parent", stopImmediatePropagation: true})
   _triggerOnListenAt() {
     __firedEvents._triggerOnListenAt = true;
   }
@@ -470,3 +503,4 @@ class MultipleHandlers extends EventMap(HTMLElement) {
   }
 }
 customElements.define("event-map-multiple-handlers", MultipleHandlers);
+