+++ packages/EventMap/CHANGELOG.md
@@ -4,6 +4,44 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.7.1] 2021-09-22
+### Updated
+- dependencies and audit
+
+
+## [0.7.0] 2021-09-22
+### Breaking Change
+- Moved decorators to a decorators file
+
+## [0.6.1] - 2021-09-10
+### Updated
+- Documentation 
+
+## [0.6.0] - 2021-09-04
+### Updated
+- Documentation
+
+## [0.5.0] - 2021-09-03
+### Added
+- Event stop propagation options
+
+
+## [0.4.0] - 2021-08-12
+### Added
+- Check and warning for calling middleware method more than once
+
+## [0.3.7] - 2021-08-07
+### Added
+- Added collapsed option to logging
+
+## [0.3.6] - 2021-08-05
+### Fixed
+- Handler binding error
+
+## [0.3.5] - 2021-08-05
+### Fixed
+- Updates for package publishing
+
 ## [0.3.1] - 2021-08-03
 ### Changed
 - Extracted methods for readability